+++ Src/xmllogger.cpp
@@ -174,7 +174,7 @@ void XmlLogger::writeToLogHPpath(const std::list<Node> &hppath)
     }
 }
 
-void XmlLogger::writeToLogSummary(unsigned int numberofsteps, unsigned int nodescreated, double length, double time, double cellSize)
+void XmlLogger::writeToLogSummary(unsigned int numberofsteps, unsigned int nodescreated, float length, double time, double cellSize)
 {
     if (loglevel == CN_LP_LEVEL_NOPE_WORD)
         return;
@@ -185,7 +185,7 @@ void XmlLogger::writeToLogSummary(unsigned int numberofsteps, unsigned int nodes
     element->SetAttribute(CNS_TAG_ATTR_NUMOFSTEPS, numberofsteps);
     element->SetAttribute(CNS_TAG_ATTR_NODESCREATED, nodescreated);
     element->SetAttribute(CNS_TAG_ATTR_LENGTH, length);
-    element->SetAttribute(CNS_TAG_ATTR_LENGTH_SCALED, length*cellSize);
+    element->SetAttribute(CNS_TAG_ATTR_LENGTH_SCALED, length*float(cellSize));
     element->SetAttribute(CNS_TAG_ATTR_TIME, std::to_string(time).c_str());
 }
